requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.2
pydantic-settings==2.1.0
python-dotenv==1.0.0
cohere==4.32
requests==2.31.0
python-linkedin-v2==0.8.3
facebook-sdk==3.1.0
redis==5.0.1
jinja2==3.1.2
python-jose[cryptography]==3.3.0
email-validator==2.1.0.post1 